+++ packages/dropdown_button2/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
-- Upgrade minimum required Flutter SDK version to 3.32.0
+- Upgrade minimum required Flutter SDK version to 3.32.0.
+- Add `errorBuilder` support for DropdownButtonFormField2 [Flutter core].
 
 ## 3.0.0